requirements.txt
streamlit>=1.32.0
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
Pillow>=10.0.0
pandas>=2.0.0
matplotlib>=3.7.0
numpy>=1.24.0
opencv-python-headless>=4.8.0